ci: 收紧多平台 Windows/Linux x86/ARM 锁版1.95.0 单编译 (feat/v2.1-impl, no merge to master)
- 矩阵 Windows/Linux x86_64/aarch64 四组合, x86跑test ARM仅check, 覆盖编译
- 锁版 rust 1.95.0/node22.18/pnpm10.28/actions pin, rust-cache复用
- 单编译 fmt->clippy->test, 严 -D warnings, 补注释说明
- 修 clippy: manual_ok_err/collapsible_if/type_complexity/unnecessary_cast

+++ tests/performance/tests/configure_large.rs
@@ -32,7 +32,7 @@ async fn bench(n: usize) -> u128 {
     for _ in 0..20 {
         let start = Instant::now();
         let _ = conn.configure(1, vec![make_task(n, "t")]).await.unwrap();
-        samples.push(start.elapsed().as_millis() as u128);
+        samples.push(start.elapsed().as_millis());
     }
     samples.sort_unstable();
     let p95 = samples[(0.95 * (samples.len() as f64 - 1.0)).round() as usize];
